QA: Ensure survey code loaded before testing if survey is visible
So after managing to replicate browser test failures locally
when running them against beta labs I managed to get them consistently
passing with this minimal change.

It seems testing for ext.quicksurvey.init is not enough as this itself loads
the code for the survey and the code to view it. Since the former is prone to
change we test for the latter as an indication it is okay to run the test.

No sleep needed! Let's put this bug to bed!!!

Bug: T113534
Change-Id: Iffbdd4926cb9041b0c42280e5acd020ddabb952a