Avoid unstubbing wgUser when an external survey is defined
Use content language to load the survey URL so that the user is
not loaded. This will break l10n of the survey URL (if that was
the intention) but retain wiki admins' ability to change it per-site.
If localization is needed, it should be reintroduced in a way
that is consistent with the ResourceLoader contract (no user-
dependency outside MakeGlobalVariablesScript).

Also fix the logic which would have considered protocol-relative
URLs insecure but any URL with 'https' anywhere in it secure.

Bug: T127870
Change-Id: Id1956d139a2c06d1e33d8a6d62d92b815a54a8fe