Don't run bucketing scenarios against Beta Cluster
The bucketing scenarios require that the "internal-" and
"external-example-survey" surveys have non-zero coverage. However, the
coverage of all surveys on the Beta Cluster was set to 0 in If681dfa3 to
reduce the noise on the Mobile 2G performance dashboard [0].

Extract the bucketing scenarios into a separate feature that isn't run
against the Beta Cluster but is still run during integration (per
change).

[0]: https://grafana.wikimedia.org/dashboard/db/mobile-2g

Bug: T159739
Change-Id: Ib30d39db9652460cdae942554130bb19bdc0ef48