init: Improve dnt code with idiomatic patterns
The navigator object is a global variable and should be used as
such. The indirection of the window proxy makes the code less
easy to read (because it's uncommon to use this way), and also
decreases the surface area for tools such as ESLint (and their
plugins in IDEs) to detect mistakes when misspelling the global
(it would be detected as undefined).

window.navigator cannot be undefined in any supported context
to execute this code.

Also adjust the comment to refer to EventLogging instead of an
unrelated npm package. What matters is that this code is in sync
with how EL determines DNT, so that we don't show the survey
when its result would be ignored, and we do show the survey
in all case where it isn't.

Change-Id: If067159ab5f249efc64ff601e05ae0382d78691d