Remove indirection of 'views' subnamespace
Remove the dead code at the bottom of utils.js for creating
`mw.extQuickSurveys` and `mw.extQuickSurveys.views`. This is an
anti-pattern that I believe originated from code I regretfully
wrote for Moodbar very early on in MW-JS experiments. At the time
I did not have confidence in how my code executed and defensively
redefined every property "just in case". This doesn't make sense
because the order of execution is well-defined. The files execute
in a reliable order. If not, there are bigger problems than this
property missing. In this case, the object is defined by lib.js.

For the others, just attach directly to the extension' main host
object. The name of a class should be stable and separate from
the bundle name (module).

Soon this will be redundant with module.exports.

Change-Id: I106848383470a5a87c35bbda0cc74bab787e4048