Remove redundant wgQuickSurveysRequireHttps from JS startup data
This feature was added in 2015 and on by default.

To support Beta Cluster, where at the time we did not yet have HTTPS
by default, the feature was made configurable and disabled in
Beta (wmf-config: 5dfe0b39c5948).

Given we now have HTTPS there by default as well, this no longer
needs to be made configurable especially through the legacy means
of adding config bytes to the startup payload on all pages.

If we need this again in the future, it should be exported by
bundling the data with the module that needs it instead of
in the criticial path through the startup bundle (as it was).

I thought of simply removing the conditional and letting true
be the default, which I thought was intended. But, it seem the
feature is (as of 1f2af9e6ffc9) both disabled by default and
in Beta Cluster. Which means it's actually dead code that has
a static and unused default value for its configuration exported,
so instead I'm commenting it out as first step to keep behaviour
as it is.

Bug: T219342
Change-Id: I386c0287acc33e847a3af2cda9822ddcd5a9afb9