Add freeform text to internal surveys
The configuration file for internal surveys has changed. If
"freeformTextLabel" option is set to a message key, the internal survey
will render a free form text input and a submit button after the answer
buttons.

Because of this [1] bug, radio buttons were not used, although, their
use may have been more elegant.

To keep changes to a minimum, no user input validation is done on the
front end. Data has to be cleaned up on the back end.

An alert is shown when the user doesn't select a button, and doesn't
enter their own answer into the free form input. There's no limit on the
number of characters the user can input. Arbitrary limits don't work
well because we cannot be sure of the total number of characters to be
sent to the EventLogging backend beforehand because the configuration
options may change thus increasing or decreasing the total. This means
there's a slight chance that some users' responses won't be recorded if
they are too long.

Also, the existing eslint warnings have been fixed.

[1] https://phabricator.wikimedia.org/T177617

Bug: T225042
Change-Id: If80f0e32a04295d6e4120d7ff4d897f1fd56b480