Depend on EventLogging
For better or worse, QuickSurveys uses EventLogging as its storage
mechanism. However, this dependency only made clear in the documentation
[0].

If the ext.eventLogging module hasn't loaded before the survey has
loaded, then mw.eventLog.logEvent won't exist. However, because we guard
the usage of that method, the survey is still interactive but silently
drops data. Depending on the ext.eventLogging module prevents this
scenario.

[0] https://www.mediawiki.org/wiki/Extension:QuickSurveys#Prerequisites

Bug: T220627
Change-Id: Ic20d97e3af9a3ab9ac6fd6be81c302951d2c24a4