Implement zero-cost conditional load of lib
The init module is added to the output of every mainspace article
unconditionally. We rely on the init module to load the rest of the
frontend codebase conditionally. We do this because we don't want to
wait for the mainspace article cache to invalidate in order to enable a
survey. Currently, however, the lib and views modules are loaded as
dependecies of the init module.

This change minimizes the performance cost of having no enabled surveys
defined by: merging the .lib and .views modules to reduce the size of
the ResourceLoader module manifest; and redefining the init module as a
subclass of ResourceLoadModule, which depends on the .lib module if
there are enabled surveys defined.

Supporting changes:

- Create the QuickSurveys.Config and QuickSurveys.EnabledSurveys
  services, which are accessed throughout the codebase

Bug: T213459
Change-Id: I924feb06d4ca62069fc36c6111185fab9860ab39