[tests] Stop test interference by extra survey
The browser tests for external surveys were written assuming that only
one internal and one external survey were enabled. This assumption was
invalidated by the introduction of the freeform textbox tests.

This change allows that assumption to hold but stops the tests for the
freeform textbox interfering by reducing the coverage of that survey to
0. The survey won't be shown unless requested using the ?quickSurvey
query parameter.

This approach is the same as for the "drink-survey" survey.

Change-Id: I93f36b31a4b0b79ae4db7e15acb7ce3bda1d4a2a