Use mw.util.debounce instead of $.debounce
$.debounce was deprecated in favour of OO.ui.debounce in If1c165d4.
However, mw.util.debounce was added in If370b959.

mediawiki.util is likely to be loaded on all pageviews as it's a
dependency of the mediawiki.Title, mediawiki.page.ready, and
mediawiki.api ResourceLoader modules.

More worryingly, QuickSurveys didn't depend on the
jquery.throttle-debounce ResourceLoader module. If this module isn't
loaded, then a TypeError will be thrown and the QuickSurveysInitiation
instrumentation won't work.

Change-Id: I39f1c3679fc0dad34cde3fd77d628f9cf45ceefd