Partially revert I1aecbf77
The introduction of a skipFunction for the .lib module increases the
cost of conditional loading of the module for case where there are no
enabled surveys defined, which is the case for the majority of the time
that the extension is enabled. This is because the skipFunction is
bundled with the ResourceLoader module manifest.

This change reverts only the addition of the skipFunction for the .lib
module in I1aecbf77. However, while it is the case that Do Not Track is
an outgoing technology [0], it makes sense to continue to check whether
the user has enabled DNT to remain consistent with other codebases.

[0] https://phabricator.wikimedia.org/T213459#5802809

Bug: T213459
Change-Id: Ibda5101ce007baf7e59e08e61865be1b87f970a7