Regression: Disable impression logging correctly
I39f1c367 introduced a regression where survey impressions are logged
not once but whenever the user scrolls. There was a pronounced increase
in the number of QuickSurveyInitiation events logged after the change
was deployed [0].

$.throttle/debounce returns wrapper functions that jQuery will treat the
same as the original callback [1]. mw.util.debounce doesn't. So when the
original logSurveyImpression function is passed to $.off, jQuery doesn't
recognize it and silently fails.

Fix this by introducing a new helper function that returns a promise
when 50% of an element has intersected with the viewport once, which:

- Allows us to remove the logSurveyImpression global state, since all
  required state is local to new helper function

- Allows clients to be agnostic to future changes to how we detect
  whether a survey has been seen, e.g. by using the IntersectionObserver
  API

[0] https://grafana.wikimedia.org/d/000000018/eventlogging-schema?orgId=1&var-schema=QuickSurveyInitiation&from=1578960000000&to=1580255999000
[1] https://gerrit.wikimedia.org/g/mediawiki/core/+/c033c6262858404aa2557eee7a31ae2ef2a6bb69/resources/lib/jquery.throttle-debounce/jquery.ba-throttle-debounce.js#186

Change-Id: Ied65793e987bab8054bbfe2f3dde9e8c6d8dac9b