Remove duplicate $name property from subclasses
The $name property already exists in the parent class. No need to
store it twice.

Note that InternalSurvey was not even populating it's duplicate
$name property. It was always null. This bug was introduced in 2015
in patch I99ffb2c as part of T119367. Luckily this didn't broke the
class. The + operator prefers elements from the first array. The
"name" element from the second – now broken – array was always
ignored.

Bug: T119367
Change-Id: I1cd329ebf53f9bde7849895e8edaf6bab29ef4f1