Remove Do Not Track support
This is the direction that all Wikimedia modules are headed, because
data protections are extended to all users regardless of their headers.

DNT users will be included in survey sampling from this point on.

DEPLOYMENT NOTE: An explanation of the change must be announced
publicly, including date stamps for production deployment.

Bug: T254224
Change-Id: I5a8573651f0a2761547438b3f878c318e9487047