Disable and log surveys with invalid configuration
Catch the validation exception and log at error level.  Invalid
surveys are filtered out the same as if they were disabled.  Before
this patch, invalid survey configuration would crash the site when
loading any page.

This will be important for softening migration to future survey spec
changes.  Now it's safe to roll back code (stopping at this patch),
even if configuration has evolved and includes breaking changes.

Preserves logic which defaults the `enabled` to false, but might it
be better to report a validation error when this key is missing (in a
future patch)?

Bug: T253112
Change-Id: I1d8391fd0189fe06bceac8f404bb4a554d92420b