Multiple-answer layout
The core QuickSurvey widget has an extensible design, leveraged in
this patch.  QuickSurvey becomes an abstract base class and the
unique part of the current `internal` survey implementation is
extracted into a `SingleAnswerSurvey` subclass.  A new `layout` field
becomes available to survey configuration, in the future becoming
mandatory.

This allows for another variation of internal survey,
`layout: multiple-answer`, which lays out an arbitrary number of
as a list of checkboxes.  The selected answers' message keys are
joined by "," when submitting the response.

The extension's minor version is bumped to reflect the new
configuration fields.

DEPLOYMENT NOTE:
Any active or new survey configurations should include
"layout: single-answer", as in I3396f925f9136, but during a migration
period this will be made the default layout.  This means that code
and config changes are safe to deploy or revert in any order.

Bug: T255130
Change-Id: I87f49ccdce04994b190b1538d9fbc0e7a0b15f38