Simplify factory with null coalesce
Note that this slightly changes the logic, so that an empty string
is accepted as valid.  My reasoning for why this is fine is that the
validation intends to check for missing fields and shouldn't be in
the business of testing every possible edge case.  For example, we
don't detect misspelled message keys.

Change-Id: Ifef5a80548a13a85502bb10cfbbcc1a3e5cc7364