Drop broken QUnit test that works only in Chrome
As far as I can see this stopped working just recently because it
contains the hard-coded assumption that this test can only ever be
run in Chrome. But who knows what user agent our QUnit test runner
uses?

Mocking navigator.userAgent is not possible. The property is
read-only.

The code in lib.js would need to be rewritten to accept
navigator.userAgent as a parameter to make it testable. Personally
I don't think this is worth it. Instead I suggest to remove the
test case.

Bug: T372596
Change-Id: I0f85521af61793e27d94a08eb634ece317d4b84e