Exclude QuickSurveys from scaling with text size
As a temporary workaround for inconsistencies caused by partial
font mode integration, this change prevents the QuickSurveys UI
component in the content area from scaling with the Vector 2022
text size switcher.
- Apply `.vector-feature-custom-font-size-clientpref--excluded`
to the root element of QuickSurveys.
- Explicitly set font-size and line-height on `.ext-quick-survey-panel`
to `@font-size-medium` and `@line-height-medium`.
- Add a comment noting that these changes are temporary and
should be removed once font modes are fully integrated in Vector.

Bug: T391890
Change-Id: I512eee9179d91157825e66bda747b0d86db39800